handle errors in delete somethings

+++ src/controllers/renters/DELETE.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { Renter } from "../../models/Renter";
+import { Contract } from "../../models/Contract";
 
 type Renter = {
   name: string;
@@ -20,6 +21,12 @@ type Renter = {
 export const deleteRenter = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const contracts = await Contract.findAll({ where: { renterId: id } });
+    if (contracts.length > 0) {
+      return res
+        .status(410)
+        .json({ message: "Cannot delete renter with contracts" });
+    }
     await Renter.destroy({ where: { id } });
     res.sendStatus(204);
   } catch (error: unknown) {
